Fix doctors rendering issue
